Separated the ranging protocol in another task
Breaks the usb communication
