+++ Makefile
@@ -31,6 +31,7 @@ HALS+=i2c_ex
 OBJS+=src/main.o
 OBJS+=src/usb_device.o src/usbd_cdc_if.o src/usbd_desc.o src/lps25h.o src/led.o
 OBJS+=src/cfg.o src/usbcomm.o src/test_support.o src/production_test.o
+OBJS+=src/uwb.o
 
 HALS+=gpio rcc cortex i2c pcd dma pcd_ex rcc_ex spi uart
 OBJS+=$(foreach mod, $(HALS), $(HAL_ROOT)/Src/stm32$(CPU)xx_hal_$(mod).o)
