+++ inc/uwb.h
@@ -1,0 +1,49 @@
+/*
+ *    ||          ____  _ __
+ * +------+      / __ )(_) /_______________ _____  ___
+ * | 0xBC |     / __  / / __/ ___/ ___/ __ `/_  / / _ \
+ * +------+    / /_/ / / /_/ /__/ /  / /_/ / / /_/  __/
+ *  ||  ||    /_____/_/\__/\___/_/   \__,_/ /___/\___/
+ *
+ * LPS node firmware.
+ *
+ * Copyright 2016, Bitcraze AB
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Foobar is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Foobar.  If not, see <http://www.gnu.org/licenses/>.
+ */
+/* uwb.h: Uwb radio implementation */
+
+#ifndef __UWB_H__
+#define __UWB_H__
+
+#include <stdbool.h>
+#include <stdint.h>
+#include "cfg.h"
+
+#define MAX_ANCHORS 6
+
+struct uwbConfig_s {
+  CfgMode mode;
+  uint8_t address[8];
+  uint8_t anchorListSize;
+  uint8_t anchors[MAX_ANCHORS];
+};
+
+void uwbInit();
+bool uwbTest();
+void uwbStart();
+char * uwbStrError();
+struct uwbConfig_s * uwbGetConfig();
+
+#endif //__UWB_H__
