+++ src/dwOps.c
@@ -34,8 +34,6 @@ extern SPI_HandleTypeDef hspi1;
 // #define DEBUG_SPI
 
 #define DWM_IRQn EXTI0_1_IRQn
-#define DWM_IRQ_PIN GPIO_PIN_0
-
 
 static dwDevice_t *dev;
 
@@ -45,28 +43,6 @@ void dwOpsInit(dwDevice_t *device)
   dev = device;
 
   NVIC_EnableIRQ(DWM_IRQn);
-}
-
-static int checkIrq()
-{
-  return HAL_GPIO_ReadPin(GPIOB, GPIO_PIN_0);
-}
-
-void HAL_GPIO_EXTI_Callback(uint16_t GPIO_Pin)
-{
-  switch (GPIO_Pin) {
-    case DWM_IRQ_PIN:
-      do{
-          dwHandleInterrupt(dev);
-      } while(checkIrq() != 0); //while IRS line active (ARM can only do edge sensitive interrupts)
-      HAL_NVIC_ClearPendingIRQ(DWM_IRQn);
-      break;
-    case GPIO_PIN_12:
-      //instance_notify_DW1000_inIDLE(1);
-      break;
-    default:
-      break;
-  }
 }
 
 // Aligned buffer of 128bytes
